dom/indexedDB/test/test_setVersion.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">

function testSteps()
{
  const READ_WRITE = Components.interfaces.nsIIDBTransaction.READ_WRITE;

  const name = window.location.pathname;
  const description = "My Test Database";

  let request = moz_indexedDB.open(name, description);
  request.onerror = errorHandler;
  request.onsuccess = grabEventAndContinueHandler;
  let event = yield;

  let db = event.result;

  // Check default state.
  is(db.version, "", "Correct default version for a new database.");

  const versions = [
    "1.0",
    "",
  ];

  for (let i = 0; i < versions.length; i++) {
    let version = versions[i];

    request = db.setVersion(version);
    request.onerror = errorHandler;
    request.onsuccess = grabEventAndContinueHandler;
    event = yield;

    is(db.version, version, "Database version number updated correctly");
    is(event.transaction.mode, READ_WRITE, "Not FULL_LOCK");
  }

  SimpleTest.executeSoon(function() { testGenerator.next(); });
  yield;

  let objectStoreNumber = 0;
  let callbackCount = 0;
  let lastObjectStore = 0;

  for (; objectStoreNumber < 20;) {
    request = db.createObjectStore(objectStoreNumber++, "");
    request.onerror = errorHandler;
    request.onsuccess = function(event) {
      is(event.transaction.mode, READ_WRITE, "Not FULL_LOCK");
      callbackCount++;
    }
  }

  let continueCreating = true;
  while (continueCreating) {
    request = db.createObjectStore(objectStoreNumber++, "");
    request.onerror = errorHandler;
    request.onsuccess = function(event) {
      is(event.transaction.mode, READ_WRITE, "Not FULL_LOCK");
      callbackCount++;
      if (parseInt(event.result.name) == 21) {
        ok(objectStoreNumber > callbackCount, "more outstanding creates");
        lastObjectStore = objectStoreNumber;

        setTimeout(function() { continueCreating = false; }, 1000);

        request = db.setVersion("11");
        request.onerror = errorHandler;
        request.onsuccess = function() {
          is(event.transaction.mode, READ_WRITE, "Not FULL_LOCK");
          is(callbackCount, lastObjectStore,
             "Version callback happened before other creates");
        }
      }
    }
    setTimeout(function() { testGenerator.next(); }, 10);
    yield;
  }

  while (callbackCount < objectStoreNumber) {
    setTimeout(function() { testGenerator.next(); }, 10);
    yield;
  }

  ok(callbackCount > lastObjectStore, "More callbacks after version");

  finishTest();
  yield;
}
  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
